requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
numexpr==2.10.1
numpy==2.2.1
outcome==1.3.0.post0
pandas==2.2.3
pip==24.2
platformdirs==4.3.6
pycparser==2.21
pygame==2.5.2
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
selenium==4.16.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
termcolor==2.5.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.28.0
Werkzeug==3.1.3
wheel==0.44.0
win-inet-pton==1.1.0
wsproto==1.2.0
